Refactor ID Types in Controllers and Repositories for Consistency
- Updated all relevant controllers and repositories to use Integer instead of Long for ID parameters, ensuring consistency across the application.
- Adjusted service methods and DTOs accordingly to reflect the change in ID types, enhancing type safety and reducing potential errors.
- Improved overall code clarity and maintainability by standardizing ID handling throughout the application.

+++ spring-boot/greenmagic/backend/src/main/java/org/sortoutinnovation/greenmagic/controller/AuthController.java
@@ -195,7 +195,7 @@ public class AuthController {
      */
     @PutMapping("/users/{userId}/password")
     public ResponseEntity<?> changePassword(
-            @PathVariable Long userId,
+            @PathVariable Integer userId,
             @Valid @RequestBody PasswordChangeRequest passwordChangeRequest) {
         try {
             UserResponseDto user = userService.updatePassword(
@@ -224,7 +224,7 @@ public class AuthController {
      */
     @PutMapping("/admin/users/{userId}/reset-password")
     public ResponseEntity<?> resetPassword(
-            @PathVariable Long userId,
+            @PathVariable Integer userId,
             @Valid @RequestBody PasswordResetRequest passwordResetRequest) {
         try {
             UserResponseDto user = userService.resetPassword(
@@ -291,7 +291,7 @@ public class AuthController {
      * @return update result
      */
     @PostMapping("/debug/make-admin/{userId}")
-    public ResponseEntity<?> makeUserAdmin(@PathVariable Long userId) {
+    public ResponseEntity<?> makeUserAdmin(@PathVariable Integer userId) {
         try {
             UserResponseDto updatedUser = userService.updateUserRole(userId, "ADMIN");
