+++ spring-boot/greenmagic/backend/src/main/java/org/sortoutinnovation/greenmagic/repository/CartItemRepository.java
@@ -10,20 +10,20 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CartItemRepository extends JpaRepository<CartItem, Long> {
+public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
     
     @Query("SELECT ci FROM CartItem ci WHERE ci.cart.cartId = :cartId")
-    List<CartItem> findByCartId(@Param("cartId") Long cartId);
+    List<CartItem> findByCartId(@Param("cartId") Integer cartId);
     
     @Query("SELECT ci FROM CartItem ci WHERE ci.cart.user.userId = :userId")
     List<CartItem> findByUserId(@Param("userId") Long userId);
     
     @Query("SELECT ci FROM CartItem ci WHERE ci.cart.cartId = :cartId AND ci.product.productId = :productId")
-    Optional<CartItem> findByCartIdAndProductId(@Param("cartId") Long cartId, @Param("productId") Long productId);
+    Optional<CartItem> findByCartIdAndProductId(@Param("cartId") Integer cartId, @Param("productId") Integer productId);
     
     @Query("SELECT ci FROM CartItem ci WHERE ci.product.productId = :productId")
-    List<CartItem> findByProductId(@Param("productId") Long productId);
+    List<CartItem> findByProductId(@Param("productId") Integer productId);
     
     @Query("SELECT COUNT(ci) FROM CartItem ci WHERE ci.cart.cartId = :cartId")
-    long countByCartId(@Param("cartId") Long cartId);
+    long countByCartId(@Param("cartId") Integer cartId);
 }
